+++ src/kernel/interupts/pit.h
@@ -1,6 +1,8 @@
 #pragma once
 
-extern volatile int pitInteruptsTriggered;
+#include <stdint.h>
+
+extern volatile uint64_t pitInteruptsTriggered;
 extern volatile int shouldSchedule;
 
 extern void pit_isr();
